refactor: reorganize axios sdk in a module
